Loaded data from supab into store successfully
