+++ src/routes/+page.svelte
@@ -2,6 +2,7 @@
     import Navbar from "$lib/NavBar/NavBar.svelte"
     import MainView from "$lib/MidView.svelte"
 
+    console.log("\n---New Instance:---\n")
 </script>
 <main>
     <Navbar/>
